refactor(davinci): split P0-11 profiler and export files under the length gate
